requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.5.0
datasets>=1.5.0
Pillow>=8.0.0

